Add files via upload
There will be three types of queries to the database

insert a planet with name  and entrance fee  into the database
increase the entrance fee for all planets between  by 
return the entrance fee for a planet name  from the database
Input Format:
The first line contains n the number of queries to be made to the database.
The next n lines contain queries of the following 3 types
1. a, k insert a planet with name a and entrance fee k into the database
2. a, b, k increase the entrance fee for all planets between a, b by k 
3. a return the entrance fee for a planet name a from the database
Constraints
n<=1000
a , b <= 10
0<k<100000

The entrance fee for any planet will always be less than 2^30

YOU MUST USE 2-3 TREE DATA STRUCTURES PROVIDED IN CLASS

As discussed in class, every node in the tree should have a value field. To avoid time-outs, all queries, including range updates (type 2 queries), should run in time O(nlogn) .

For each query of type 3 print the entrance fee of the planet in a new line.